significantly improve not-working conversion from graphalg core -> relalg. make it compatible with mlir-opt
